+++ packages/transaction-messages/src/index.ts
@@ -2,6 +2,7 @@ export * from './blockhash';
 export * from './codecs';
 export * from './compile'; // TODO later: can probably delete this at the end
 export * from './create-transaction-message';
+export * from './decompile-message';
 export * from './durable-nonce';
 export * from './fee-payer';
 export * from './instructions';
