+++ packages/transactions/src/index.ts
@@ -1,7 +1,6 @@
 export * from './blockhash';
 export * from './compilable-transaction';
 export * from './create-transaction';
-export * from './decompile-transaction';
 export * from './durable-nonce';
 export * from './fee-payer';
 export * from './instructions';
